requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
azure-core==1.24.2
    # via
    #   azure-identity
    #   azure-storage-blob
    #   azure-storage-file-datalake
    #   msrest
azure-functions==1.11.2
    # via KNMI-FA (setup.py)
azure-identity==1.10.0
    # via KNMI-FA (setup.py)
azure-storage-blob==12.13.0
    # via azure-storage-file-datalake
azure-storage-file-datalake==12.8.0
    # via KNMI-FA (setup.py)
certifi==2022.6.15
    # via
    #   msrest
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.0
    # via requests
cryptography==37.0.4
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
idna==3.3
    # via requests
isodate==0.6.1
    # via msrest
msal==1.18.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.0.0
    # via azure-identity
msrest==0.7.1
    # via
    #   azure-storage-blob
    #   azure-storage-file-datalake
oauthlib==3.2.0
    # via requests-oauthlib
portalocker==2.5.1
    # via msal-extensions
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.4.0
    # via msal
requests==2.28.1
    # via
    #   KNMI-FA (setup.py)
    #   azure-core
    #   msal
    #   msrest
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via msrest
six==1.16.0
    # via
    #   azure-core
    #   azure-identity
    #   isodate
typing-extensions==4.3.0
    # via azure-core
urllib3==1.26.10
    # via requests
